+++ src/portfolio/capital.rs
@@ -11,7 +11,7 @@ pub trait CapitalHandlers {
     fn decrease_capital<T>(&mut self, amount: f64, point: T)
     where
         T: Into<Option<NaiveDateTime>>;
-    fn get_capital(&self) -> f64;
+    fn available_capital(&self) -> f64;
 }
 
 impl CapitalHandlers for Portfolio {
@@ -29,7 +29,7 @@ impl CapitalHandlers for Portfolio {
         self.capital_ts.decrement(amount, point);
     }
 
-    fn get_capital(&self) -> f64 {
+    fn available_capital(&self) -> f64 {
         self.capital_ts.get_last_value()
     }
 }
@@ -42,7 +42,7 @@ mod tests {
 
         let mut portfolio = Portfolio::new(100.0, 100.0, None);
         portfolio.increase_capital(10.0, None);
-        assert_eq!(portfolio.get_capital(), 110.0);
+        assert_eq!(portfolio.available_capital(), 110.0);
     }
 
     #[test]
@@ -51,7 +51,7 @@ mod tests {
 
         let mut portfolio = Portfolio::new(100.0, 100.0, None);
         portfolio.decrease_capital(10.0, None);
-        assert_eq!(portfolio.get_capital(), 90.0);
+        assert_eq!(portfolio.available_capital(), 90.0);
     }
 
     #[test]
@@ -59,12 +59,12 @@ mod tests {
         use super::*;
 
         let mut portfolio = Portfolio::new(100.0, 100.0, None);
-        assert_eq!(portfolio.get_capital(), 100.0);
+        assert_eq!(portfolio.available_capital(), 100.0);
 
         portfolio.increase_capital(10.0, None);
-        assert_eq!(portfolio.get_capital(), 110.0);
+        assert_eq!(portfolio.available_capital(), 110.0);
 
         portfolio.decrease_capital(10.0, None);
-        assert_eq!(portfolio.get_capital(), 100.0);
+        assert_eq!(portfolio.available_capital(), 100.0);
     }
 }