+++ CHANGELOG.md
@@ -1,5 +1,37 @@
 # CHANGELOG
 
+## v0.2.1
+
+### New Features
+
+- Implements a simple backtesting engine
+- Use `rust_decimal` crate for accurate decimal arithmetic. This has not fully been implemented yet.
+
+### Strategy Improvements
+
+- Implement `Strategy.get_combined_signals` (untested)
+
+### Portfolio Improvements
+
+- Add performance metrics to `Portfolio`
+- Add functions which propose `FutureTrades` - `generate_buy_opt()` and `generate_sell_opt()`
+- Mark `Portfolio.is_rate_profitable` as deprecated. Usage must be removed from `Engine`.
+- Add `PortfolioArgs` for `Portfolio` initialization
+- Rename `get_capital` to `available_capital` for better readability
+- Remove `Persistence` module. This will be replaced with a database implementation in the future.
+
+### Indicator Improvements
+
+- Add `Err` return values to `BBands.process_graph_for_new_candles()`
+- Formally test `BBands.Gprocess_graph_for_new_candles()`
+
+### Utility Improvements
+
+- Create util for extracting `Vec<Candle>` from a dataframe
+- Create util for extracting `Vec<Side>` from a dataframe
+
+---
+
 ## v0.2.0
 
 This release adds significant performance to `BBands`, and changes the API for `Strategy` and `Indicator` in preparation for a backtesting implementation.